requirements.txt
cachetools==5.5.2 ; python_version >= "3.12"
certifi==2025.1.31 ; python_version >= "3.12"
charset-normalizer==3.4.1 ; python_version >= "3.12"
google-api-core==2.24.2 ; python_version >= "3.12"
google-api-python-client==2.166.0 ; python_version >= "3.12"
google-auth-httplib2==0.2.0 ; python_version >= "3.12"
google-auth-oauthlib==1.2.1 ; python_version >= "3.12"
google-auth==2.38.0 ; python_version >= "3.12"
googleapis-common-protos==1.69.2 ; python_version >= "3.12"
httplib2==0.22.0 ; python_version >= "3.12"
idna==3.10 ; python_version >= "3.12"
numpy==2.2.4 ; python_version >= "3.12"
oauthlib==3.2.2 ; python_version >= "3.12"
pandas-stubs==2.2.3.250308 ; python_version >= "3.12"
pandas==2.2.3 ; python_version >= "3.12"
proto-plus==1.26.1 ; python_version >= "3.12"
protobuf==6.30.2 ; python_version >= "3.12"
pyasn1-modules==0.4.2 ; python_version >= "3.12"
pyasn1==0.6.1 ; python_version >= "3.12"
pymarc==5.2.3 ; python_version >= "3.12"
pyparsing==3.2.3 ; python_version >= "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.12"
pytz==2025.2 ; python_version >= "3.12"
requests-oauthlib==2.0.0 ; python_version >= "3.12"
requests==2.32.3 ; python_version >= "3.12"
rsa==4.2 ; python_version >= "3.12"
six==1.17.0 ; python_version >= "3.12"
types-pytz==2025.2.0.20250326 ; python_version >= "3.12"
types-requests==2.32.0.20250328 ; python_version >= "3.12"
tzdata==2025.2 ; python_version >= "3.12"
uritemplate==4.1.1 ; python_version >= "3.12"
urllib3==2.3.0 ; python_version >= "3.12"
